Moving constants into separate files
